split up now incorrectly named spatial_scene_part_data file
